requirements.txt
prometheus-client
python-jenkins
requests
python-dotenv
gunicorn